+++ Sources/Display/Win32/input_device_provider_win32mouse.h
@@ -54,10 +54,10 @@ public:
 	InputDevice::Type get_type() const { return InputDevice::pointer; }
 
 	/// \brief Returns the x position of the device.
-	int get_x() const;
+	float get_x() const;
 
 	/// \brief Returns the y position of the device.
-	int get_y() const;
+	float get_y() const;
 
 	/// \brief Returns true if the passed key code is down for this device.
 	bool get_keycode(int keycode) const;
@@ -98,7 +98,7 @@ public:
 	}
 
 	/// \brief Sets the position of the device.
-	void set_position(int x, int y);
+	void set_position(float x, float y);
 
 /// \}
 /// \name Implementation
